# server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "七日记（Note4Seven）API",
    "version": "0.1.0",
    "description": "Generated route inventory with shared runtime contracts. Examples must use synthetic data."
  },
  "servers": [
    {
      "url": "https://staging.invalid",
      "description": "Staging placeholder; inject the real domain outside source control"
    }
  ],
  "paths": {
...
    "/api/v1/modules/{moduleId}/streak-rewards": {
      "get": {
        "operationId": "get_api_v1_modules_moduleId_streak_rewards",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/modules/{moduleId}/streak-rewards",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "post_api_v1_modules_moduleId_streak_rewards",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/modules/{moduleId}/streak-rewards",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}/streak-rewards/{rewardRuleId}": {
      "delete": {
        "operationId": "delete_api_v1_modules_moduleId_streak_rewards_rewardRuleId",
        "tags": [
          "v1"
        ],
        "summary": "DELETE /api/v1/modules/{moduleId}/streak-rewards/{rewardRuleId}",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rewardRuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}/streak-rewards/{rewardRuleId}/preview": {
      "post": {
        "operationId": "post_api_v1_modules_moduleId_streak_rewards_rewardRuleId_preview",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/modules/{moduleId}/streak-rewards/{rewardRuleId}/preview",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rewardRuleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}/streak-rewards/pending": {
      "get": {
        "operationId": "get_api_v1_modules_moduleId_streak_rewards_pending",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/modules/{moduleId}/streak-rewards/pending",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}/streak-rewards/received": {
      "get": {
        "operationId": "get_api_v1_modules_moduleId_streak_rewards_received",
        "tags": [
          "v1"
        ],
        "summary": "GET /api/v1/modules/{moduleId}/streak-rewards/received",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/streak-reward-draws/{drawId}/redeem": {
      "post": {
        "operationId": "post_api_v1_streak_reward_draws_drawId_redeem",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/streak-reward-draws/{drawId}/redeem",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "drawId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/streak-reward-draws/{drawId}/reveal": {
      "post": {
        "operationId": "post_api_v1_streak_reward_draws_drawId_reveal",
        "tags": [
          "v1"
        ],
        "summary": "POST /api/v1/streak-reward-draws/{drawId}/reveal",
        "x-source": "src/routes/streak-rewards.ts",
        "parameters": [
          {
            "name": "drawId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Quota exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "requestId",
          "serverTime"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId",
          "serverTime"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
    }
  }
}